Use Boost::Asio::static_thread_pool to limit the number of concurrent queries (#1103)
So far, the number of concurrent queries running in QLever was limited by first acquiring a semaphore and then scheduling the query on a completely new thread using handwritten glue code. Now those queries are simply scheduled onto a thread pool that automatically handles the number of concurrent queries.

+++ src/ServerMain.cpp
@@ -126,8 +126,8 @@ int main(int argc, char** argv) {
             << qlever::version::GitShortHash() << EMPH_OFF << std::endl;
 
   try {
-    Server server(port, static_cast<int>(numSimultaneousQueries),
-                  memoryMaxSizeGb, std::move(accessToken), !noPatternTrick);
+    Server server(port, numSimultaneousQueries, memoryMaxSizeGb,
+                  std::move(accessToken), !noPatternTrick);
     server.run(indexBasename, text, !noPatterns, !onlyPsoAndPosPermutations);
   } catch (const std::exception& e) {
     // This code should never be reached as all exceptions should be handled
